:wrench: Replace old atmosphere plugin
