--- plugins/csp-advanced-atmospheres/src/logger.cpp
@@ -1,23 +1,0 @@
-////////////////////////////////////////////////////////////////////////////////////////////////////
-//                               This file is part of CosmoScout VR                               //
-////////////////////////////////////////////////////////////////////////////////////////////////////
-
-// SPDX-FileCopyrightText: German Aerospace Center (DLR) <cosmoscout@dlr.de>
-// SPDX-License-Identifier: MIT
-
-#include "logger.hpp"
-
-#include "../../../src/cs-utils/logger.hpp"
-
-namespace csp::atmospheres {
-
-////////////////////////////////////////////////////////////////////////////////////////////////////
-
-spdlog::logger& logger() {
-  static auto logger = cs::utils::createLogger("csp-atmospheres");
-  return *logger;
-}
-
-////////////////////////////////////////////////////////////////////////////////////////////////////
-
-} // namespace csp::atmospheres
