+++ plugins/csp-atmospheres/CMakeLists.txt
@@ -13,23 +13,16 @@ endif()
 
 # build plugin -------------------------------------------------------------------------------------
 
-file(GLOB SOURCE_FILES src/*.cpp)
-
-set(TEST_FILES)
-
-if (COSMOSCOUT_UNIT_TESTS)
-  file(GLOB TEST_FILES test/*.cpp)
-endif()
+file(GLOB SOURCE_FILES src/*.cpp src/*/*.cpp src/*/*/*.cpp src/*/*/*/*.cpp)
 
 # Resoucre files and header files are only added in order to make them available in your IDE.
 file(GLOB HEADER_FILES src/*.hpp)
-file(GLOB_RECURSE RESOUCRE_FILES gui/* textures/*)
+file(GLOB_RECURSE RESOUCRE_FILES shaders/*)
 
 add_library(csp-atmospheres SHARED
   ${SOURCE_FILES}
   ${HEADER_FILES}
   ${RESOUCRE_FILES}
-  ${TEST_FILES}
 )
 
 target_link_libraries(csp-atmospheres
@@ -51,9 +44,4 @@ source_group(TREE "${CMAKE_CURRENT_SOURCE_DIR}" FILES
 # install plugin -----------------------------------------------------------------------------------
 
 install(TARGETS   csp-atmospheres  DESTINATION "share/plugins")
-install(DIRECTORY "textures"       DESTINATION "share/resources")
-install(DIRECTORY "gui"            DESTINATION "share/resources")
-
-if (COSMOSCOUT_UNIT_TESTS)
-  install(DIRECTORY "test/reference" DESTINATION "bin/test")
-endif()
+install(DIRECTORY "shaders"        DESTINATION "share/resources")
