+++ plugins/csp-atmospheres/gui/atmospheres_settings.html
@@ -46,9 +46,10 @@ SPDX-License-Identifier: MIT
 
 <div class="row">
   <div class="col-5">
-    Quality
+    Cloud Altitude
   </div>
   <div class="col-7">
-    <div data-callback="atmosphere.setQuality"></div>
+    <div data-callback="atmosphere.setCloudAltitude">
+    </div>
   </div>
 </div>
