+++ plugins/csp-atmospheres/gui/js/csp-atmospheres.js
@@ -19,8 +19,8 @@
      * @inheritDoc
      */
     init() {
-      CosmoScout.gui.initSlider("atmosphere.setQuality", 1, 30, 1, [7]);
-      CosmoScout.gui.initSlider("atmosphere.setWaterLevel", -2, 2, 0.01, [0]);
+      CosmoScout.gui.initSlider("atmosphere.setCloudAltitude", 0, 10000, 10, [2000]);
+      CosmoScout.gui.initSlider("atmosphere.setWaterLevel", -10000, 10000, 10, [0]);
     }
   }
