+++ plugins/csp-atmospheres/shaders/csp-atmospheres/atmosphere.frag
