+++ plugins/csp-atmospheres/shaders/csp-atmospheres/models/bruneton/definitions.glsl
