+++ plugins/csp-atmospheres/shaders/csp-atmospheres/models/bruneton/functions.glsl
