+++ plugins/csp-atmospheres/shaders/csp-atmospheres/models/cosmoscout/model.glsl
