+++ plugins/csp-atmospheres/src/ModelBase.hpp
