+++ plugins/csp-atmospheres/src/models/bruneton/Model.cpp
