+++ plugins/csp-atmospheres/src/models/bruneton/Model.hpp
