+++ plugins/csp-atmospheres/src/models/bruneton/internal/Model.cpp
