+++ plugins/csp-atmospheres/src/models/bruneton/internal/constants.hpp
