+++ plugins/csp-atmospheres/src/models/cosmoscout/Model.cpp
