+++ plugins/csp-atmospheres/src/models/cosmoscout/Model.hpp
