--- plugins/csp-atmospheres/test/reference/csp-atmospheres-Atmosphere-01.png.license
@@ -1,3 +1,0 @@
-SPDX-FileCopyrightText: 2022 German Aerospace Center (DLR) <cosmoscout@dlr.de>
-
-SPDX-License-Identifier: CC0-1.0
