--- plugins/csp-atmospheres/textures/earth-clouds.jpg.license
@@ -1,3 +1,0 @@
-SPDX-FileCopyrightText: 2022 NASA (https://visibleearth.nasa.gov/view.php?id=57747)
-
-SPDX-License-Identifier: CC0-1.0
